Refactor the build process
New specialists are now added with a form and and a script.
The rest of the codebase has been refactored to make these
easier to use.
