+++ .github/workflows/new-entry-branches.yml
@@ -1,0 +1,137 @@
+name: Create branches for new entries
+
+on:
+  issues:
+    types: [opened]
+
+jobs:
+  comment:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+          ref: main
+      - uses: actions/github-script@v6
+        env:
+          GIT_AUTHOR_NAME: "Automatic updater [bot]"
+          GIT_AUTHOR_EMAIL: "sleepdiary-bot@pileofstuff.org"
+          GIT_COMMITTER_NAME: "Automatic updater [bot]"
+          GIT_COMMITTER_EMAIL: "sleepdiary-bot@pileofstuff.org"
+          AUTHOR: ${{ github.event.issue.user.login }}
+          TITLE: ${{ github.event.issue.title }}
+          BODY: ${{ github.event.issue.body }}
+        with:
+          script: |
+            const author = process.env.AUTHOR;
+            const title = process.env.TITLE;
+            const body  = process.env.BODY;
+            const issue_number = ${{ github.event.issue.number }};
+
+            const branch_name = 'issue-'+issue_number;
+
+            const fs = require('fs');
+            const child_process = require('child_process');
+
+            title.replace(
+                /^New (specialist|software):/,
+                (_,entity_type) => body.replace(/```(.*)```/s, (_,text) => {
+                    const entry = JSON.parse(text);
+                    let todo_list = '';
+
+                    entry.id = new Date().getTime().toString();
+                    entry.last_updated = new Date().toISOString().split("T")[0];
+
+                    [ 'forms', 'reports' ].forEach(
+                        fr => entry[fr].forEach( (doc,n) => {
+                            if ( doc.url ) {
+                                doc.gallery = [
+                                    {
+                                        thumb: `/resources/thumbs/${doc.name||entry.name}.jpg`,
+                                        url: doc.url,
+                                    }
+                                ];
+                                delete doc.url;
+                            } else {
+                                doc.gallery = "\uE000";
+                                todo_list += `- [ ] Create a gallery for ${fr}[${n}]\n`;
+                            }
+                            doc["\uE001"] = "(line used for internal processing)";
+                            todo_list += `- [ ] Fill in extra information for \`${fr}[${n}]\`\n`;
+                        })
+                    );
+
+                    todo_list += `- [ ] Remove TODO items
+            - [ ] Merge the issue branch into the main branch
+            - [ ] [Generate docs from resources](../../docs/actions/workflows/generate-from-resources.yml)
+            - [ ] Any last checks/process improvements`
+
+                    let entry_text = JSON.stringify(entry,null,'  ');
+                    entry_text = (
+                        entry_text
+                            .replace(
+                                /"\uE000"/g,
+                                `[
+            #          TODO: fill in the values below:
+            #          {
+            #            "url": "...",
+            #            "thumb": "..."
+            #          },
+                    ]`
+                            )
+                            .replace(
+                                /.*\uE001.*/g,
+                                `#        TODO: fill in the values below:
+            #        "layout": "calendar", # or "graph" or "questionnaire"
+            #        "start_page": 1, # or the index of the relevant page in the PDF
+            #        "total_pages": # e.g. 1 or "variable"
+            #        "page_duration": # how blocks are separated, e.g. "14 days" or "2 weeks"
+            #        "start_time": # e.g. "Noon" or "6pm"
+            #        "inbed_marker": # e.g. "&darr;" or "(none)"
+            #        "outofbed_marker": # e.g. "&uarr;" or "(none)"
+            #        "sleep_marker": # e.g. ▬ or ▮
+            #        "events": [
+            #          {
+            #            "key": # e.g. "C"
+            #            "value": # e.g. "coffee, cola, or tea"
+            #          },
+            #             ...
+            #         ]`
+                            )
+                            .replace(
+                                /.*\uE003.*/g,
+                                "# TODO: check and push this new entry:"
+                            )
+                            .replace( /\n*$/, "\n" )
+                    );
+
+                    const filename = `entities/${entity_type}/${Math.floor(new Date().getTime()/1000)}-${entry.name.toLowerCase().replace(/[^a-z0-9]+/g,'-')}.json`;
+                    fs.writeFileSync(filename, entry_text);
+
+                    const spawnSync = require("child_process").spawnSync;
+
+                    spawnSync('git', ['add',filename]);
+                    spawnSync('git', ['commit',filename,'-m','First draft for issue #' + issue_number + ', thanks to @' + author]);
+
+                    spawnSync('git', ['push','origin','HEAD:'+branch_name]);
+
+                    let message = `Thanks for contributing!
+
+            Here's [an auto-generated first draft](https://github.com/${context.repo.owner}/${context.repo.repo}/tree/${branch_name}/${filename}) based on your information.
+
+            If you'd like to edit this draft yourself, please [fork this repository](https://github.com/${context.repo.owner}/${context.repo.repo}/fork) then [edit \`${filename}\`](https://github.com/${context.repo.owner}/${context.repo.repo}/edit/${branch_name}/${filename}).  Just let us know the repository we need to look at.
+
+            Otherwise, someone should come along in a day or two and finish off the last few jobs:
+
+            ` + todo_list;
+
+                    github.rest.issues.createComment({
+                        issue_number,
+                        owner: context.repo.owner,
+                        repo: context.repo.repo,
+                        body: message
+                    });
+
+                })
+            );
+
