+++ Makefile
@@ -1,4 +1,4 @@
 build: entities.json thumbs
 
-entities.json: compile-entities.js entities/*.yaml software-thumbs/*
+entities.json: bin/compile-entities.js entities/*/*.json software-thumbs/*
 	./$<
