+++ entities/specialist/1651137204-northshore-sleep-center.json
@@ -1,0 +1,132 @@
+{
+  "id": "1651395453580",
+  "name": "NorthShore Sleep Center",
+  "admin_contact": "",
+  "description": "Offers state-of-the-art, comprehensive care for patients suffering from disorders including sleep apnea, restless legs syndrome, night terrors, snoring, insomnia and narcolepsy.",
+  "forms": [
+    {
+      "instructions_to_find": "1. Go to https://www.northshore.org/sleep-center/\n2. Click on \"Patient Forms\" in the left menu\n3. Click on \"Two Week Sleep Log\"\n",
+      "gallery": [
+        {
+          "thumb": "/resources/thumbs/NorthShore Sleep Center.jpg",
+          "url": "https://www.northshore.org/globalassets/sleepcenter/2weeksleeplog.pdf"
+        }
+      ],
+      "before_first": "used_maybe",
+      "between_appointments": "unused",
+      "inbed_marker": "&darr;",
+      "layout": "calendar",
+      "name": "Two Week Sleep Log",
+      "outofbed_marker": "&uarr;",
+      "page_duration": "14 days",
+      "sharing_status": "found online",
+      "sleep_marker": "▬",
+      "start_page": 1,
+      "start_time": "6pm",
+      "total_pages": 1
+    }
+  ],
+  "last_updated": "2022-04-09",
+  "locations": [
+    {
+      "address": "9977 Woods Drive\nSkokie,\nIL 60077",
+      "gps": [
+        "42.061500035075575",
+        "-87.76042044195773"
+      ],
+      "name": "Old Orchard Sleep Center"
+    },
+    {
+      "address": "2151 Waukegan Road\nBannockburn,\nIL 60015",
+      "fax": "847.236.9637",
+      "gps": [
+        "42.192501074063905",
+        "-87.85818768489362"
+      ],
+      "name": "Bannockburn Sleep Center",
+      "tel": "847.236.1300"
+    },
+    {
+      "address": "2650 Ridge Avenue,\nEvanston,\nIL 60201",
+      "gps": [
+        "42.06573038320172",
+        "-87.68442312438462"
+      ],
+      "name": "Evanston Hospital",
+      "tel": "847.570.2000",
+      "url": "https://www.northshore.org/locations/our-hospitals/evanston-hospital/"
+    },
+    {
+      "address": "2100 Pfingsten Road,\nGlenview\nIL 60026",
+      "gps": [
+        "42.092992193320626",
+        "-87.85421200035263"
+      ],
+      "name": "Glenbrook Hospital",
+      "tel": "847.657.5800"
+    },
+    {
+      "address": "7900 Rollins Rd.\nSuite B1300\nGurnee,\nIL 60031",
+      "fax": "847.236.9637",
+      "gps": [
+        "42.384974224740624",
+        "-87.99647169737688"
+      ],
+      "name": "Gurnee Medical Office Building",
+      "tel": "847.236.1300"
+    },
+    {
+      "address": "777 Park Avenue West,\nHighland Park,\nIL 60035",
+      "gps": [
+        "42.19142955063223",
+        "-87.8084486035028"
+      ],
+      "name": "Highland Park Hospital",
+      "tel": "847.432.8000",
+      "url": "https://www.northshore.org/locations/our-hospitals/highland-park-hospital/"
+    },
+    {
+      "address": "920 Milwaukee Ave.\nSuite 2100\nLincolnshire,\nIL 60069",
+      "fax": "847.570.2984",
+      "gps": [
+        "42.183534386843206",
+        "-87.92753459645732"
+      ],
+      "name": "Lincolnshire Medical Group Office building",
+      "tel": "847.663.8200"
+    },
+    {
+      "address": "9600 Gross Point Road,\nSkokie,\nIL 60076",
+      "gps": [
+        "42.05672098158611",
+        "-87.74040640008995"
+      ],
+      "name": "Skokie Hospital",
+      "tel": "847.677.9600",
+      "url": "https://www.northshore.org/locations/our-hospitals/skokie-hospital/"
+    }
+  ],
+  "registration_required": false,
+  "appointment_types": [
+    "in-person"
+  ],
+  "booking_required": true,
+  "booking_url": "https://www.northshore.org/OpenScheduling/PhysicianDirectSchedule.aspx?pid=7374",
+  "booking_tel": "847.663.8200",
+  "booking_fax": "",
+  "booking_other": "[Richard John Munson, M.D.](https://www.northshore.org/apps/findadoctor/physicians/richard-john-munson) specialises in sleep-related conditions",
+  "before_first": "you may be asked to fill out some of [their patient forms](https://www.northshore.org/sleep-center/patients/)",
+  "between_appointments": "you may be referred for an outpatient sleep study at a [sleep lab](https://www.northshore.org/sleep-center/#sleep-center).",
+  "outcomes": [
+    "diagnose",
+    "treat",
+    "refer"
+  ],
+  "outcome_other": "",
+  "procedure_type": "researched",
+  "referral_types": [
+    "direct"
+  ],
+  "specialist_type": "physician",
+  "url": "https://www.northshore.org/sleep-center/"
+}
